+++ src/main/java/it/gov/pagopa/pu/debtpositions/connector/classification/config/ClassificationApiClientConfig.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.debtpositions.connector.classification.config;
 
-import it.gov.pagopa.pu.debtpositions.config.ApiClientConfig;
+import it.gov.pagopa.pu.debtpositions.config.rest.ApiClientConfig;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
