+++ src/main/java/it/gov/pagopa/pu/debtpositions/connector/organization/config/OrganizationApiClientConfig.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.debtpositions.connector.organization.config;
 
-import it.gov.pagopa.pu.debtpositions.config.ApiClientConfig;
+import it.gov.pagopa.pu.debtpositions.config.rest.ApiClientConfig;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
